Skip tests for docs folders (#281)
Our mandatory checks will fail if there are docs only changes. The
helm-chart-ci.yaml skips these files, so the workflows never run, and
the mandatory checks are stuck waiting for results.

This adds a parallel workflow that runs fake tests for docs changes. So
that docs only changes can pass the required status checks. It uses
`paths-ignore` instead of `paths`. This is documented
[here](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-protected-branches/troubleshooting-required-status-checks#handling-skipped-but-required-checks).

---------

Signed-off-by: Faisal Memon <fymemon@yahoo.com>
Co-authored-by: Marco Franssen <marco.franssen@gmail.com>