Fix bundle role/rolebinding naming conflict
When you install a primary and a secondary deployment in the same
Kubernetes cluster, use the same release name (spire), and configure
the primary to give the bundle to the secondary, both charts want to
own the same role/rolebinding name. This patch adds the release
namespace into the names to prevent conflict.

Signed-off-by: Kevin Fox <Kevin.Fox@pnnl.gov>