Removed pull_request edited event from ci workflow
Edited event is only useful if you want a workflow to act on a PR description or title or labels and such, we are not doing that in this workflow.

Removing the edited type from the event doesn't unnecesarily run the workflow again when title, description and such are changed.

Signed-off-by: Marco Franssen <marco.franssen@gmail.com>