Fix bundle role/rolebinding naming conflict
When you install a primary and a secondary deployment in the same Kubernetes
cluster, use the same release name (spire), and configure the primary to
give the bundle to the secondary, both releases want to own the same
role/rolebinding name. This patch uses the same name as the configmap
to ensure uniqueness.

Signed-off-by: Kevin Fox <Kevin.Fox@pnnl.gov>